+++ joulescope_ui/devices/jsdrv/jsdrv_wrapper.py
@@ -102,9 +102,6 @@ class JsdrvWrapper:
         self._stream_buffers[mem_id] = b
         unique_id = f'JsdrvStreamBuffer:{mem_id:03d}'
         self.pubsub.register(b, unique_id=unique_id)
-        topic = get_topic_name(unique_id)
-        for device in self.devices.values():
-            self.pubsub.publish(f'{topic}/actions/!device_add', device)
 
     def on_action_mem_remove(self, value):
         self._log.info('mem remove %s', value)